Add support for multiple inventory files
This was a lot more work than I had hoped. In the end, it wasn't so
hard, it was more a matter of being able to figure out how to get `jq`
to convert the JSON output from `ansible-inventory` in the way I
desired.

One thing I'm still not please about _at all_ is the performance.
Completion is now fairly slow, and the fact that we are using separate
commands for the hosts and groups contribute to this fact.

$ time ansible all --list-hosts 2> /dev/null | tail +2 > /dev/null && time ansible-inventory --list | jq -r "keys | .[]" > /dev/null

real    0m0.214s
user    0m0.187s
sys 0m0.030s

real    0m0.460s
user    0m0.395s
sys 0m0.068s

However, I'll try to live with it for now. It's still much better than
before, since I was previously unable to have a comma-separated list of
inventory files listed in the `ansible.cfg` file.