Cache ansible inventory data (#2)
Invoking `ansible` and `ansible-inventory` on each completion incurs a
delay of at least 300 + 300 ms. In my experience, the actual latency
is even worse (more like 700-800 ms). Now, this might seem a tolerable
latency for some, but for something that you execute interactively,
while typing a command, it's far too much for my taste.

The naïve approach to fix this here is to just cache this output in a
`.txt` file and `cat` this file on subsequent invocations. It improves
the performance significantly, going down to ~3 ms on my local machine.

Downside: cache invalidation. Whenever a host or group is added/removed
to/from the Ansible inventory, the cached content would have to be
invalidated. A simple approach would be to just delete it in a `cron`
job, but in my experience it's very likely that you want to be able to
execute one or more Ansible commands pretty much immediately after
updating the inventory. I think I will probably just make an
`ansible_clear_cache` alias locally to delete the file and execute it
when I need it. :-)

A more advanced approach would be to check the timestamp on the file and
only allow it to be used until a given TTL has expired. (Yes, we are
going into overingineering now.) Even better: kick of a background
process to update the cached version anytime completion has been
performed. This would require some thinking to get it done right,
though.